% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cannonball-package.R
\docType{package}
\name{cannonball-package}
\alias{cannonball}
\alias{cannonball-package}
\title{cannonball: Tools for Teaching Statistics}
\description{
User-friendly functions for teaching introductory and intermediate statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/janhove/cannonball}
  \item \url{https://janhove.github.io/cannonball/}
  \item Report bugs at \url{https://github.com/janhove/cannonball/issues}
}

}
\author{
\strong{Maintainer}: Jan Vanhove \email{janvanhove@gmail.com} (\href{https://orcid.org/0000-0002-4607-4836}{ORCID})

Authors:
\itemize{
  \item Jan Vanhove \email{janvanhove@gmail.com} (\href{https://orcid.org/0000-0002-4607-4836}{ORCID})
}

}
\keyword{internal}
