% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal.R
\name{reveal}
\alias{reveal}
\title{Reveal the true data's position in a diagnostic parade}
\usage{
reveal(parade)
}
\arguments{
\item{parade}{The name of an object generated using the `parade()`
or `parade_summary()` function.}
}
\description{
The title says it all.
}
\examples{
# Fit model
m <- lm(mpg ~ disp, data = mtcars)

# Generate parade
my_parade <- parade(m)

# Linearity check
lin_plot(my_parade)

# Reveal
reveal(my_parade)
}
