% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R
\name{test_statistics}
\alias{test_statistics}
\alias{mean_diff}
\alias{stud_mean_diff}
\alias{median_diff}
\alias{rank_sum}
\alias{prob_super}
\alias{ks_stat}
\title{Test statistics for two-group comparisons}
\usage{
mean_diff(outcome, treatment_idx)

stud_mean_diff(outcome, treatment_idx)

median_diff(outcome, treatment_idx)

rank_sum(outcome, treatment_idx)

prob_super(outcome, treatment_idx)

ks_stat(outcome, treatment_idx)
}
\arguments{
\item{outcome}{A vector of numeric outcomes.}

\item{treatment_idx}{The indices of the observations in the treatment group.}
}
\value{
The test statistic for the two-group comparison.
}
\description{
These functions compute test statistics that can be used
in a randomisation test for a two-group comparison.

\code{mean_diff()} computes the mean difference.

\code{stud_mean_diff()} computes the studentised mean difference.
This is identical to the t-statistic used in Welch's t-test.

\code{median_diff()} computes the difference between the medians.

\code{rank_sum()} computes the rank sum in the treatment group.
This is the test statistic used in the Mann-Whitney-Wilcoxon test.

\code{prob_super()} computes the probability of superiority, i.e.,
the probability that a randomly chosen unit in the treatment group
has a higher outcome value than a randomly chosen unit in the control group.
The resulting test is equivalent to when using \code{rank_sum()}.

\code{ks_stat()} computes the Kolmogorov-Smirnov test statistic, i.e.,
the supremum of the difference between the ecdfs of the outcome in the
treatment and control groups.
}
\examples{
x <- c(seq(-4, 5), 2*seq(-3, 4))
g <- rep(c(0, 1), times = c(10, 8))
idx <- which(g == 1)
mean_diff(x, idx)
stud_mean_diff(x, idx)
median_diff(x, idx)
rank_sum(x, idx)
prob_super(x, idx)
ks_stat(x, idx)

}
