% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkthrough_blocking.R
\name{walkthrough_blocking}
\alias{walkthrough_blocking}
\title{Walkthrough blocking on a covariate}
\usage{
walkthrough_blocking(
  n = 10,
  diff = 0,
  sd = 1,
  rho = 0.8,
  showdata = FALSE,
  M = NULL
)
}
\arguments{
\item{n}{The number of data points per group.}

\item{diff}{The boost that participants in the intervention group receive.}

\item{sd}{The standard deviation of the normal distributions from which the data are drawn.}

\item{rho}{The correlation between the covariate and the outcome (pre-intervention) in the population.}

\item{showdata}{Do you want to output a dataframe containing the plotted data (\code{TRUE})
or not (\code{FALSE}, default)?}

\item{M}{\code{NULL} (default) when using exhaustive randomisation testing; else set to the number of Monte Carlo runs desired.}
}
\description{
This function produces a step-by-step demonstration
of how researchers can 'block' on a continuous covariate
and how they can analyse the data of a randomised block design.
}
\details{
Data are generated from a normal distribution with the requested
standard deviation; a covariate is also generated. The data points are then
grouped in pairs based on their covariate scores. Within each pair, the data points
are then randomly assigned to the control or intervention group.
Data points in the intervention group receive a boost
as specified by 'diff'. Finally, a significance test is ran on the data.
This significance test is a randomisation test using the mean difference as
the test statistic. The p-value reported is a two-sided one.
}
\examples{
\dontrun{
walkthrough_blocking(n = 12, diff = 0.2, sd = 1.3, rho = 0.8)

# Save data and double check results
dat <- walkthrough_blocking(n = 12, diff = 0.2, sd = 1.3, rho = 0.8, showdata = TRUE)
anova(lm(score ~ factor(block) + group, data = dat))
}
}
\keyword{blocking,}
\keyword{control}
\keyword{p-value,}
\keyword{significance}
\keyword{test,}
\keyword{variable}
