% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cannonball-package.R
\docType{package}
\name{cannonball-package}
\alias{cannonball}
\alias{cannonball-package}
\title{cannonball: Tools for Teaching Statistics}
\description{
User-friendly functions for teaching introductory and intermediate statistics. Julian `Cannonball` Adderley is one of the author's favourite saxophonists and was a consummate jazz educator.
}
\author{
\strong{Maintainer}: Jan Vanhove \email{jan.vanhove@unifr.ch}

}
\keyword{internal}
