% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plot.R
\name{diagnostic_plots}
\alias{diagnostic_plots}
\alias{lin_plot}
\alias{var_plot}
\alias{norm_qq}
\alias{norm_hist}
\title{Draw diagnostic parades}
\usage{
lin_plot(parade)

var_plot(parade)

norm_qq(parade)

norm_hist(parade, bins = 30)
}
\arguments{
\item{parade}{The name of an object generated using `parade()`.
For `var_plot()`, objects generated using `resid_summary()` are also accepted.}

\item{bins}{How many bins should the histograms contain? Defaults to 30.}
}
\description{
These are short-hand functions to quickly draw
diagnostic parades.

`lin_plot()` plots the models' residuals against their corresponding fitted values.
This is useful for checking the linearity assumption.

`var_plot()` plots the models' absolute residuals against their corresponding fitted values
if it is fed an object generated using `parade()`. If it is fed an object
generated using `resid_summary()`, it plots the sample standard deviation of the residuals
per cell.
This is useful for checking the constant-variance assumption.

`norm_qq()` and `norm_hist()` plot normal quantile-quantile plots and histograms of the
models' residuals, respectively.
This is useful for checking the normality assumption.
}
\examples{
# A simple regression model
m <- lm(mpg ~ disp, data = mtcars)

# Generate parade and check linearity
my_parade <- parade(m)
lin_plot(my_parade)
reveal(my_parade)

# Regenerate parade and check constant variance
my_parade <- parade(m)
var_plot(my_parade)
reveal(my_parade)

# Regenerate parade and check normality
my_parade <- parade(m)
norm_qq(my_parade)
norm_hist(my_parade)
norm_hist(my_parade, bins = 10)
reveal(my_parade)
}
\keyword{dogs}
