% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parade.R
\name{parade}
\alias{parade}
\title{Generate dataset for a diagnostic parade}
\usage{
parade(model, full_data = NULL)
}
\arguments{
\item{model}{The name of the statistical model you want to diagnose.
Currently only lm(), gam() (from the `mgcv` package) and lmer() (from the `lme4` package) models are supported.
For the lmer() models, only residual diagnostics are supported; support for BLUP ('random effects') diagnostics
is still lacking.}

\item{full_data}{By default, the output will only include variables that are part of the model.
If you want to include all the variables that are present in the dataframe
on which the model was fitted, supply this dataframe's name to full_data.}
}
\description{
This function generates a parade (= 'lineup' in the `nullabor` package)
that hides the observations, fitted values, and residuals of a statistical model you
want to diagnose among the observations, fitted values, and residuals of 19 similar models
that were fitted on simulated outcome data. The 19 sets of simulated outcome data are
generated from the original model so that this model's assumptions are literally true
for the simulated data. The 'tibble' (dataframe) created by this can be used to draw
20 panels of diagnostic plots (see examples).
}
\examples{
# A simple regression model
m <- lm(mpg ~ disp, data = mtcars)

# Generate parade and check linearity
my_parade <- parade(m)
my_parade
lin_plot(my_parade)
reveal(my_parade)

# Regenerate parade and check constant variance
my_parade <- parade(m)
var_plot(my_parade)
reveal(my_parade)

# Regenerate parade and check normality
my_parade <- parade(m)
norm_qq(my_parade)
norm_hist(my_parade)
norm_hist(my_parade, bins = 10)
reveal(my_parade)
}
\keyword{assumptions,}
\keyword{diagnostics,}
\keyword{lineup}
\keyword{model}
\keyword{protocol}
