% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal.R
\name{reveal}
\alias{reveal}
\title{Reveal the true data's position in a diagnostic parade}
\usage{
reveal(parade)
}
\arguments{
\item{parade}{The name of an object generated using the \code{\link{parade}}
or \code{\link{parade_summary}} function.}
}
\description{
This function reveals the position of the true data in a parade.
}
\examples{
# Fit model
m <- lm(mpg ~ disp, data = mtcars)

# Generate parade
my_parade <- parade(m)

# Linearity check
lin_plot(my_parade)

# Reveal
reveal(my_parade)
}
