% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkthrough_p.R
\name{walkthrough_p}
\alias{walkthrough_p}
\title{Walkthrough p-values}
\usage{
walkthrough_p(n = 10, diff = 0, sd = 1, showdata = FALSE, M = NULL)
}
\arguments{
\item{n}{The number of data points per group.}

\item{diff}{The boost that participants in the intervention group receive.}

\item{sd}{The standard deviation of the normal distributions from which the data are drawn.}

\item{showdata}{Do you want to output a dataframe containing the plotted data (\code{TRUE})
or not (\code{FALSE}, default)?}

\item{M}{\code{NULL} (default) when using exhaustive randomisation testing; else set to the number of Monte Carlo runs desired.}
}
\description{
This function produces a step-by-step demonstration
of a significance test for a two-group comparison.
}
\details{
Data are generated from a normal distribution with the requested
standard deviation. Then, the data points are randomly assigned to two
equal-sized groups. Data points in the intervention group receive a uniform boost
as specified by \code{diff}. Finally, a significance test is run on the data.
This significance test is a randomisation test using the mean difference as
the test statistic. The p-value reported is a two-sided one.

If \code{n} is larger than 9 and \code{M} is not specified, \code{M} is set to 48620.
}
\examples{
\dontrun{
walkthrough_p(n = 12, diff = 0.2, sd = 1.3)

# Save data and double check results using Welch t-test
dat <- walkthrough_p(n = 10, diff = 0.2, sd = 2, showdata = TRUE)
t.test(score ~ group, data = dat)
}
}
\keyword{p-value}
\keyword{significance}
\keyword{test,}
